# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Comic Canvas API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/feedback/pair": {
      "post": {
        "tags": [
          "feedback"
        ],
        "summary": "Feedback Pair",
        "operationId": "feedback_pair_feedback_pair_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PairIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Feedback Pair Feedback Pair Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/feedback/pick": {
      "post": {
        "tags": [
          "feedback"
        ],
        "summary": "Feedback Pick",
        "description": "User request 2026-09-20: a quick tap is a pick of the favourite of a panel's three\noptions. Writes one `image_pairs` row per option passed on, all sharing a `tap_id`, so the\npreference model can learn it as a single tap (`app.preference.retrain._load_taps`) while\nolder single-pair rows keep working. The strip switches to the pick exactly as `/pair`\ndoes.",
        "operationId": "feedback_pick_feedback_pick_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PickIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Feedback Pick Feedback Pick Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/feedback/rating": {
      "post": {
        "tags": [
          "feedback"
        ],
        "summary": "Feedback Rating",
        "description": "User request 2026-09-21: a rating (off / ok / great) of a single image. A pick says which of\nthree is best; a rating also says whether an image is good at all, so the models learn from\nboth. A changed mind is a new row (the latest per image counts). Nothing on screen changes,\nso only the learning is re-run.",
        "operationId": "feedback_rating_feedback_rating_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RatingIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Feedback Rating Feedback Rating Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/feedback/caption": {
      "post": {
        "tags": [
          "feedback"
        ],
        "summary": "Feedback Caption",
        "operationId": "feedback_caption_feedback_caption_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CaptionIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Feedback Caption Feedback Caption Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/feedback/thumb": {
      "post": {
        "tags": [
          "feedback"
        ],
        "summary": "Feedback Thumb",
        "operationId": "feedback_thumb_feedback_thumb_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ThumbIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Feedback Thumb Feedback Thumb Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/days/{date}/panels/{panel_id}/regenerate": {
      "post": {
        "tags": [
          "feedback"
        ],
        "summary": "Regenerate Panel",
        "operationId": "regenerate_panel_days__date__panels__panel_id__regenerate_post",
        "parameters": [
          {
            "name": "date",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Date"
            }
          },
          {
            "name": "panel_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Panel Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegenerateIn"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Regenerate Panel Days  Date  Panels  Panel Id  Regenerate Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/library": {
      "get": {
        "tags": [
          "library"
        ],
        "summary": "Library",
        "operationId": "library_library_get",
        "parameters": [
          {
            "name": "month",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^\\d{4}-\\d{2}$",
              "title": "Month"
            }
          },
          {
            "name": "mood",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Mood"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Library Library Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/search": {
      "get": {
        "tags": [
          "library"
        ],
        "summary": "Search",
        "operationId": "search_search_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "title": "Q"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Search Search Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CaptionIn": {
        "properties": {
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "panel_id": {
            "type": "integer",
            "title": "Panel Id"
          },
          "original": {
            "type": "string",
            "title": "Original"
          },
          "edited": {
            "type": "string",
            "title": "Edited"
          }
        },
        "type": "object",
        "required": [
          "date",
          "panel_id",
          "original",
          "edited"
        ],
        "title": "CaptionIn"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PairIn": {
        "properties": {
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "panel_id": {
            "type": "integer",
            "title": "Panel Id"
          },
          "chosen": {
            "type": "string",
            "title": "Chosen"
          },
          "rejected": {
            "type": "string",
            "title": "Rejected"
          },
          "source": {
            "type": "string",
            "enum": [
              "ab",
              "regenerate"
            ],
            "title": "Source"
          }
        },
        "type": "object",
        "required": [
          "date",
          "panel_id",
          "chosen",
          "rejected",
          "source"
        ],
        "title": "PairIn"
      },
...
      "PickIn": {
        "properties": {
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "panel_id": {
            "type": "integer",
            "title": "Panel Id"
          },
          "picked": {
            "type": "string",
            "title": "Picked"
          },
          "others": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 5,
            "minItems": 1,
            "title": "Others"
          }
        },
        "type": "object",
        "required": [
          "date",
          "panel_id",
          "picked",
          "others"
        ],
        "title": "PickIn",
        "description": "Best-of-3: the person picked `picked` as their favourite of the options shown\n(`others` are the ones they passed on). One tap, learned as the pick beating each other."
      },
      "RatingIn": {
        "properties": {
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "panel_id": {
            "type": "integer",
            "title": "Panel Id"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "rating": {
            "type": "integer",
            "maximum": 2.0,
            "minimum": 0.0,
            "title": "Rating"
          }
        },
        "type": "object",
        "required": [
          "date",
          "panel_id",
          "url",
          "rating"
        ],
        "title": "RatingIn",
        "description": "0 off, 1 ok, 2 great \u2014 for one image (usually the one in the strip)."
      },
      "RegenerateIn": {
        "properties": {
          "note": {
            "type": "string",
            "maxLength": 300,
            "title": "Note",
            "default": ""
          }
        },
        "type": "object",
        "title": "RegenerateIn"
      },
...
      "ThumbIn": {
        "properties": {
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "panel_id": {
            "type": "integer",
            "title": "Panel Id"
          },
          "value": {
            "type": "integer",
            "enum": [
              1,
              -1
            ],
            "title": "Value"
          }
        },
        "type": "object",
        "required": [
          "date",
          "panel_id",
          "value"
        ],
        "title": "ThumbIn"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}